apps/requirements.txt
# Dependências para Event Hub Kafka com OAuth
confluent-kafka>=2.3.0
azure-identity>=1.15.0
